% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMMarginal.R
\name{Marginalikelihood}
\alias{Marginalikelihood}
\title{Calculating Marginal Likelihiood or Topic Model}
\format{null}
\usage{
null
}
\arguments{
\item{wordspertopic}{Number of words per topic (integer)}

\item{LDAresult}{Results of LDA model}

\item{dtmatrix}{Document term matrix}

\item{draws}{Number of draws in MCMC}

\item{seed}{Seed number}

\item{verbose}{Print out results for each document if TRUE}
}
\value{
Return a matrix of marginal probability
Return a matrix of marginal probability (with penalization of unlikely documents)
Return a visualization of the best K
}
\description{
Calculate the marginal likelihood.
}
\examples{
library(topicmodels)
library(MCMCpack)
data("AssociatedPress")
model <- LDA(AssociatedPress[1:10,],k=5, control = list(alpha=0.1, seed=99) ,  method = 'VEM')
Marginalikelihood(10, model, AssociatedPress[1:10,],1000,99,verbose = FALSE)


with raw text data
text <- lda::lexicalize(text)
dtm <- topicmodels::ldaformat2dtm(text$document, text$vocab)
model <- LDA(dtm, k=5, control = list(alpha=0.1, seed=99) ,  method = 'VEM')
Marginalikelihood(10, model, dtm, 1000, 99,verbose = FALSE)


}
\author{
Naijia Liu
}
