% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMMarginal.R
\name{visualizeTMM}
\alias{visualizeTMM}
\title{Visualizing Marginal Likelihiood for Topic Model}
\usage{
null
}
\arguments{
\item{Kstart}{starting value of topic number}

\item{Kend}{ending value of topic number}

\item{dtmatrix}{Document term matrix}

\item{seed}{Seed number}

\item{wordspertopic}{Number of words per topic (integer)}

\item{draws}{Number of draws in MCMC}

\item{verbose}{Print out results for each document if TRUE}
}
\description{
This function visualizes the marginal likelihood value, and returns a vector of log-likelihood.
}
\examples{
library(topicmodels)
library(MCMCpack)
data("AssociatedPress")
visualizeTMM(2, 10, dtmatrix = AssociatedPress[1:5,], 99, 5, 1000, FALSE)


}
\author{
Naijia Liu
}
